Various update

+++ src/ui/common.h
@@ -7,7 +7,7 @@ void enterAltScr();
 void exitAltScr();
 void clearScr();
 void clearLine(int row, int col, int len);
-void printAt(int col, int row, const std::string &s, int color = 0, bool border = false);
+void printAt(size_t col, size_t row, const std::string &s, size_t color = 0, bool border = false);
 void refreshBuffer();
 void hideCursor();
 void showCursor();
